backend/requirements.txt
asgiref==3.2.3
Django==3.0
django-cors-headers==3.2.0
django-leaflet==0.26.0
djangorestframework==3.10.3
djangorestframework-gis==0.15
markdown==3.1.1
Pillow==6.2.1
psycopg2==2.8.4
pytz==2019.3
PyYAML==5.3b1
six==1.13.0
sqlparse==0.3.0
